Move auxv and cpuinfo parsing into arm_linux submodule
