+++ src/cpu_feature/arm_linux/test-data/macos-virtualbox-linux-x64-4850HQ.auxv
